1. Notes: Revised Note taking design to match the system theme 2. System Users: a. Upload image revisions and bug fixes b. Added delete functionality (Employee_Status = INACTIVE) c. Added password expiry date (Expiry_Date) d. Added login attempt counter (Lock_Count) 3. Clients: a. Added delete functionality b. Added validation checking during deletion (checks for existing notes and orders) 4. Revision on layout file: Added the custom file upload input

+++ app/Http/Controllers/ClientController.php
@@ -83,8 +83,17 @@ class ClientController extends BaseController
     public function deleteClient($id)
     { 
         $client = Client::findOrFail($id); 
-        $client->delete();
- 
+        $hasOrders = $client->orders()->exists();
+        if ($hasOrders) {
+            return redirect()->route('clients')->with('error', 'Cannot delete client with existing orders.');
+        } 
+
+        $hasNotes = $client->Notes()->exists(); 
+        if ($hasNotes) {
+            return redirect()->route('clients')->with('error', 'Cannot delete client with existing notes.');
+        } 
+        
+        $client->delete(); 
         return redirect()->route('clients')->with('success', 'Client deleted successfully.'); 
     }
